+++ index.js
@@ -1,95 +1,89 @@
 /**
  * Module dependencies.
  */
+'use strict'
 
-var Counter = require('passthrough-counter');
-var humanize = require('humanize-number');
-var bytes = require('bytes');
-var chalk = require('chalk');
-var co = require('co');
-
-/**
- * TTY check for dev format.
- */
-
-var isatty = process.stdout.isTTY;
+const Counter = require('passthrough-counter')
+const humanize = require('humanize-number')
+const bytes = require('bytes')
+const chalk = require('chalk')
 
 /**
  * Expose logger.
  */
 
-module.exports = dev;
+module.exports = dev
 
 /**
  * Color map.
  */
 
-var colorCodes = {
+const colorCodes = {
+  7: 'magenta',
   5: 'red',
   4: 'yellow',
   3: 'cyan',
   2: 'green',
-  1: 'green'
-};
+  1: 'green',
+  0: 'yellow'
+}
 
 /**
  * Development logger.
  */
 
-function dev(opts) {
+function dev (opts) {
   opts = opts || {};
-  opts.consoleEnabled = typeof opts.consoleEnabled === 'boolean' ? 
+  opts.consoleEnabled = typeof opts.consoleEnabled === 'boolean' ?
     opts.consoleEnabled : true;
-  opts.externalLogger = opts.externalLogger || null;
+  opts.externalLogger = typeof opts.externalLogger === 'function' ?
+    opts.externalLogger : null;
 
-  return function *logger(next) {
+  return async function logger (ctx, next) {
     // request
-    var start = Date.now();
+    const start = Date.now()
     if (opts.consoleEnabled) {
-      console.log('  ' + chalk.gray('<--')
-        + ' ' + chalk.bold('%s')
-        + ' ' + chalk.gray('%s'),
-          this.method,
-          this.originalUrl);  
-    }    
+      console.log('  ' + chalk.gray('<--') +
+        ' ' + chalk.bold('%s') +
+        ' ' + chalk.gray('%s'),
+          ctx.method,
+          ctx.originalUrl)
+    }
 
     try {
-      yield next;
+      await next()
     } catch (err) {
       // log uncaught downstream errors
-      yield log(this, start, null, err, null, opts);
-      throw err;
+      await log(ctx, start, null, err, null, opts)
+      throw err
     }
 
     // calculate the length of a streaming response
     // by intercepting the stream with a counter.
     // only necessary if a content-length header is currently not set.
-    var length = this.response.length;
-    var body = this.body;
-    var counter;
-    if (null == length && body && body.readable) {
-      this.body = body
+    const length = ctx.response.length
+    const body = ctx.body
+    let counter
+    if (length == null && body && body.readable) {
+      ctx.body = body
         .pipe(counter = Counter())
-        .on('error', this.onerror);
+        .on('error', ctx.onerror)
     }
 
     // log when the response is finished or closed,
     // whichever happens first.
-    var ctx = this;
-    var res = this.res;
+    const res = ctx.res
 
-    var onfinish = done.bind(null, 'finish');
-    var onclose = done.bind(null, 'close');
+    const onfinish = done.bind(null, 'finish')
+    const onclose = done.bind(null, 'close')
 
-    res.once('finish', onfinish);
-    res.once('close', onclose);
+    res.once('finish', onfinish)
+    res.once('close', onclose)
 
-    function done(event){
-      res.removeListener('finish', onfinish);
-      res.removeListener('close', onclose);
-      co(function*() {
-        yield log(ctx, start, counter ? counter.length : length, null, event, opts);
-      }).catch(function(err) { });      
+    function done (event) {
+      res.removeListener('finish', onfinish)
+      res.removeListener('close', onclose)
+      log(ctx, start, counter ? counter.length : length, null, event, opts)
     }
   }
 }
@@ -98,31 +92,42 @@ function dev(opts) {
  * Log helper.
  */
 
-function* log(ctx, start, len, err, event, opts) {
+async function log (ctx, start, len, err, event, opts) {
   // get the status code of the response
-  var status = err
+  const status = err
     ? (err.status || 500)
-    : (ctx.status || 404);
+    : (ctx.status || 404)
 
   // set the color of the status code;
-  var s = status / 100 | 0;
-  var color = colorCodes[s];
+  const s = status / 100 | 0
+  const color = colorCodes.hasOwnProperty(s) ? colorCodes[s] : 0
 
   // get the human readable response length
-  var length;
+  // let length
+  // if (~[204, 205, 304].indexOf(status)) {
+  //   length = ''
+  // } else if (len == null) {
+  //   length = '-'
+  // } else {
+  //   length = len
+  // }
+  let length, lenStr
   if (~[204, 205, 304].indexOf(status)) {
-    length = '';
-  } else if (null == len) {
-    length = 0;
+    length = 0
+    lenStr = ''
+  } else if (len == null) {
+    length = 0
+    lenStr = '-'
   } else {
-    length = len;
+    length = len
+    lenStr = bytes(len).toLowerCase()
   }
 
-  var upstream = err ? chalk.red('xxx')
+  const upstream = err ? chalk.red('xxx')
     : event === 'close' ? chalk.yellow('-x-')
-    : chalk.gray('-->');
+    : chalk.gray('-->')
 
-  var duration = Date.now() - start;
+  const duration = Date.now() - start
 
   if (opts.externalLogger) {
     var logObj = {
@@ -134,21 +139,21 @@ function* log(ctx, start, len, err, event, opts) {
       context: ctx,
       error: err
     };
-    yield opts.externalLogger(logObj);
+    await opts.externalLogger(logObj);
   }
 
   if (opts.consoleEnabled) {
-    console.log('  ' + upstream
-      + ' ' + chalk.bold('%s')
-      + ' ' + chalk.gray('%s')
-      + ' ' + chalk[color]('%s')
-      + ' ' + chalk.gray('%s')
-      + ' ' + chalk.gray('%s'),
+    console.log('  ' + upstream +
+      ' ' + chalk.bold('%s') +
+      ' ' + chalk.gray('%s') +
+      ' ' + chalk[color]('%s') +
+      ' ' + chalk.gray('%s') +
+      ' ' + chalk.gray('%s'),
         ctx.method,
         ctx.originalUrl,
         status,
         time(duration),
-        bytes(length));
+        lenStr)
   }
 }
 
@@ -158,9 +163,8 @@ function* log(ctx, start, len, err, event, opts) {
  * in seconds otherwise.
  */
 
-function time(delta) {
-  delta = delta < 10000
+function time (delta) {
+  return humanize(delta < 10000
     ? delta + 'ms'
-    : Math.round(delta / 1000) + 's';
-  return humanize(delta);
+    : Math.round(delta / 1000) + 's')
 }